+++ Test/Assets/Scripts/SceneButtonManager.cs
@@ -23,9 +23,6 @@ public class SceneButtonManager : MonoBehaviour
     [SerializeField] private PlayerInput playerInput;
     [SerializeField] private GameObject pauseObj;
     [SerializeField] private GameObject clearObj;
-    [SerializeField] private GameObject startUIObj;
-    [SerializeField] private GameObject canTakeFhotoObj;
-    [SerializeField] private GameObject connectingTextObj;
     
     public State CurrentState { get { return currentState; } }
     
@@ -80,44 +77,7 @@ public class SceneButtonManager : MonoBehaviour
         if (clearObj != null) {clearObj.SetActive(true);}
         Debug.Log("Game Clear:" + currentState);
     }
-
-    public void AskCanTakeOBS()
-    {
-        if(canTakeFhotoObj != null){canTakeFhotoObj.SetActive(true);}
-        Debug.Log("SetActiveTakeOBS");
-    }
-
-    public void TakeOBS()
-    {
-        if(startUIObj != null){startUIObj.SetActive(false);}
-        if(canTakeFhotoObj != null){canTakeFhotoObj.SetActive(false);}
-        if(connectingTextObj != null){connectingTextObj.SetActive(true);}
-        StartCoroutine(OBSConnection());
-        //OBSConnect();
-    }
-
-    private static async void OBSConnect()
-    {
-        Debug.Log("Host:" + RecordController.Host + "/Port:" + RecordController.Port + "/Password:" +
-                  RecordController.Password);
-        await RecordController.OBSConnect(new CancellationToken());
-        Debug.Log("OBS Conecting");
-        Debug.unityLogger.Log("OBS Connected");
-        await Task.Delay(1000);
-        RecordController.OBSRecordStart();
-    }
-    IEnumerator OBSConnection()
-    {
-        Debug.Log("Host:" + RecordController.Host + "/Port:" + RecordController.Port + "/Password:" +
-                  RecordController.Password);
-        RecordController.OBSConnect(CancellationToken.None);
-        Debug.Log("OBS Conecting");
-        //yield return new WaitUntil(RecordController.OBSIsConnected);
-        yield return new WaitForSeconds (1.0f);
-        Debug.unityLogger.Log("OBS Connected");
-        RecordController.OBSRecordStart();
-        SceneChangeMainMenu();
-    }
+    
     public void SceneChangeTitle()
     {
         InputReset();
